tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.canvas.alpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.canvas.alpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.canvas.alpha</h1>
<p class="desc">Shadows are drawn correctly for partially-transparent canvases</p>


<script>
var t = async_test("Shadows are drawn correctly for partially-transparent canvases");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var canvas2 = new OffscreenCanvas(100, 50);
  var ctx2 = canvas2.getContext('2d');
  ctx2.fillStyle = 'rgba(255, 0, 0, 0.5)';
  ctx2.fillRect(0, 0, 100, 50);

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.shadowOffsetY = 50;
  ctx.shadowColor = '#00f';
  ctx.drawImage(canvas2, 0, -50);

  _assertPixelApprox(canvas, 50,25, 127,0,127,255, 2);
  t.done();

});
</script>
